docs(m4): refresh S03 verification evidence
