+++ mods/mod-swooper-maps/src/domain/foundation/ops/index.ts
@@ -1,6 +1,9 @@
 import type { DomainOpImplementationsForContracts } from "@swooper/mapgen-core/authoring";
 import type { contracts } from "./contracts.js";
 
+import computeEraPlateMembership from "./compute-era-plate-membership/index.js";
+import computeEraTectonicFields from "./compute-era-tectonic-fields/index.js";
+import computeHotspotEvents from "./compute-hotspot-events/index.js";
 import computeCrust from "./compute-crust/index.js";
 import computeCrustEvolution from "./compute-crust-evolution/index.js";
 import computeMantleForcing from "./compute-mantle-forcing/index.js";
@@ -9,8 +12,13 @@ import computeMesh from "./compute-mesh/index.js";
 import computePlateGraph from "./compute-plate-graph/index.js";
 import computePlateMotion from "./compute-plate-motion/index.js";
 import computePlatesTensors from "./compute-plates-tensors/index.js";
+import computeSegmentEvents from "./compute-segment-events/index.js";
 import computeTectonicHistory from "./compute-tectonic-history/index.js";
+import computeTectonicHistoryRollups from "./compute-tectonic-history-rollups/index.js";
+import computeTectonicProvenance from "./compute-tectonic-provenance/index.js";
+import computeTectonicsCurrent from "./compute-tectonics-current/index.js";
 import computeTectonicSegments from "./compute-tectonic-segments/index.js";
+import computeTracerAdvection from "./compute-tracer-advection/index.js";
 
 const implementations = {
   computeMesh,
@@ -21,6 +29,14 @@ const implementations = {
   computePlateGraph,
   computePlateMotion,
   computeTectonicSegments,
+  computeEraPlateMembership,
+  computeSegmentEvents,
+  computeHotspotEvents,
+  computeEraTectonicFields,
+  computeTectonicHistoryRollups,
+  computeTectonicsCurrent,
+  computeTracerAdvection,
+  computeTectonicProvenance,
   computeTectonicHistory,
   computePlatesTensors,
 } as const satisfies DomainOpImplementationsForContracts<typeof contracts>;
@@ -28,6 +44,9 @@ const implementations = {
 export default implementations;
 
 export {
+  computeEraPlateMembership,
+  computeEraTectonicFields,
+  computeHotspotEvents,
   computeCrust,
   computeCrustEvolution,
   computeMantleForcing,
@@ -36,6 +55,11 @@ export {
   computePlateGraph,
   computePlateMotion,
   computePlatesTensors,
+  computeSegmentEvents,
   computeTectonicHistory,
+  computeTectonicHistoryRollups,
+  computeTectonicProvenance,
+  computeTectonicsCurrent,
   computeTectonicSegments,
+  computeTracerAdvection,
 };
